+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Honeycomb OpenTelemetry SDK Changelog
 * Error logging API for manually logging exceptions.
 * Package is now available on Cocoapods.
 * Add new options to enable/disable built-in auto-instrumentation.
+* Uncaught exception handler to log crashes.
 
 ### Fixes